+++ packages/anon-aadhaar-react/src/constants.ts
@@ -1,2 +1,3 @@
 export const CERTIFICATION_TUTO_URL =
   'https://anon-aadhaar-documentation.vercel.app/docs/manage-docs-versions'
+export const APP_ID = '1133947717227078484480271110087215175039824429056'