+++ packages/anon-aadhaar-react/test/pcd.test.ts
@@ -3,6 +3,7 @@ import { AnonAadhaarPCDArgs, verify, genData } from 'anon-aadhaar-pcd'
 import { assert } from 'chai'
 import { ArgumentTypeName } from '@pcd/pcd-types'
 import { proveWithWebProver } from '../src/prove'
+import { APP_ID } from '../src/constants'
 
 describe('PCD tests', function () {
   this.timeout(0)
@@ -26,6 +27,10 @@ describe('PCD tests', function () {
       base_message: {
         argumentType: ArgumentTypeName.BigInt,
         value: testData[3] + '',
+      },
+      app_id: {
+        argumentType: ArgumentTypeName.BigInt,
+        value: APP_ID,
       },
     }